#27 - moves home screen to home module
