+++ app/src/androidTest/java/br/com/rbrthmn/ui/financialcompanion/ComFinNavGraphUITest.kt
@@ -28,8 +28,8 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.navigation.compose.ComposeNavigator
 import androidx.navigation.testing.TestNavHostController
+import br.com.rbrthmn.home.ui.HOME_SCREEN_CONTENT_TEST_TAG
 import br.com.rbrthmn.ui.financialcompanion.navigation.ComFinNavGraph
-import br.com.rbrthmn.ui.financialcompanion.screens.home.HOME_SCREEN_CONTENT_TEST_TAG
 import org.junit.Test
 
 class ComFinNavGraphUITest : br.com.rbrthmn.ui.BaseUITest() {
