+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/ComFinApp.kt
@@ -39,11 +39,11 @@ import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import br.com.rbrthmn.R
+import br.com.rbrthmn.home.ui.HomeDestination
 import br.com.rbrthmn.navigation.ComFinNavigationBar
 import br.com.rbrthmn.navigation.ComFinNavigationType
 import br.com.rbrthmn.navigation.NavigationItemContent
 import br.com.rbrthmn.ui.financialcompanion.navigation.ComFinNavGraph
-import br.com.rbrthmn.ui.financialcompanion.screens.home.HomeDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.morefeatures.MoreFeaturesDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.operations.OperationsDestination
